+++ AmescoAPI/Models/Transaction.cs
@@ -6,7 +6,7 @@ namespace AmescoAPI.Models
     {
         public int TransactionId { get; set; }
         public int UserId { get; set; }
-        public int EarnedPoints { get; set; }
+        public decimal EarnedPoints { get; set; }
         public DateTime DateIssued { get; set; }
     }
 }
